requirements.txt
Flask==2.3.3
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
nltk==3.8.1
python-dateutil==2.8.2
Werkzeug==2.3.7
regex==2023.8.8
PyMuPDF
python-docx